Arregla pequeñas cosas
